Change timeout to long type

+++ c/build_all/arduino/AzureIoT/src/util/NTPClient.cpp
@@ -12,7 +12,7 @@ int NTPClient::begin()
     return _udp.begin(LOCAL_UDP_PORT);
 }
 
-uint32_t NTPClient::getEpochTime(const char* host, int port, int timeout)
+uint32_t NTPClient::getEpochTime(const char* host, int port, long timeout)
 {
     if (host == NULL || port < 1) {
         return (uint32_t)-1;
@@ -58,7 +58,7 @@ void NTPClient::sendRequest(const char* host, int port)
     _udp.endPacket();
 }
 
-int NTPClient::receiveResponse(int timeout)
+int NTPClient::receiveResponse(long timeout)
 {
     long start = millis();
     int size = 0;
