+++ c/build_all/arduino/AzureIoT/src/util/NTPClient.h
@@ -14,13 +14,13 @@ class NTPClient
     public:
         NTPClient();
         int begin();
-        uint32_t getEpochTime(const char* host, int port = NTP_PORT, int timeout = DEFAULT_NTP_TIMEOUT);
+        uint32_t getEpochTime(const char* host, int port = NTP_PORT, long timeout = DEFAULT_NTP_TIMEOUT);
         void end();
 
     private:
         void prepareRequest();
         void sendRequest(const char* host, int port);
-        int receiveResponse(int timeout);
+        int receiveResponse(long timeout);
         uint32_t parseResponse();
 
         char        _buffer[NTP_PACKET_SIZE];
